requirements.txt
picamera
pillow